Update and rename index.975ef6c8.js to index.js
